requirements/base.txt
# Archivo requirements base

django-crispy-forms==1.6.1
django-debug-toolbar==1.8.0
django-haystack==2.6.1
Django==1.11.2
markdown2==2.3.4
Pillow==4.1.1
psycopg2==2.7.1
pytz==2017.2
Whoosh==2.7.4
